Cambie el tipo de dato de double a BigDecimal

+++ App/src/main/java/ar/utn/capgemini/ecommerce/model/entities/ProductoBase.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import javax.validation.constraints.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +25,7 @@ public class ProductoBase extends EntidadPersistente {
     @NotNull
     @Positive
     @Column(name = "precioBase")
-    private double precioBase;
+    private BigDecimal precioBase;
 
     @NotNull
     @Positive
@@ -59,7 +60,7 @@ public class ProductoBase extends EntidadPersistente {
     @Column(name = "fechaUltimaModificacion", columnDefinition = "DATETIME")
     private LocalDateTime fechaUltimaModificacion;
 
-    public ProductoBase(String descripcion, double precioBase, Integer tiempoDeFabricacion, String productoBaseUrl, Categoria categoria, LocalDateTime fechaDeAlta) {
+    public ProductoBase(String descripcion, BigDecimal precioBase, Integer tiempoDeFabricacion, String productoBaseUrl, Categoria categoria, LocalDateTime fechaDeAlta) {
         this.descripcion = descripcion;
         this.precioBase = precioBase;
         this.tiempoDeFabricacion = tiempoDeFabricacion;
